# Conflicts:
#	Circle App/dev file/Events.json
#	Circle App/dev file/Settings.json

+++ src/main/java/com/example/demo/Subject.java
@@ -10,13 +10,17 @@ import java.util.ArrayList;
 
 import java.util.concurrent.ThreadLocalRandom;
 
+
+
 public class Subject extends SaveState{
 
     private String name;
+    public static String defaultName = "new folder";
 
-    public String filePath;
+    private String filePath;
     private String notesPath;
     private String pdfPath;
+    private String imagePath;
 
     public String cardPath;
     private ArrayList<CueCard> cueCardsList;
@@ -41,6 +45,7 @@ public class Subject extends SaveState{
 
         File subjectDir = new File(this.filePath + "/notes");
         File pdfDir = new File(this.filePath + "/pdfs");
+        File imgDir = new File(this.filePath + "/images");
 
         boolean success= true;//for checking if all the folders were made
 
@@ -65,12 +70,24 @@ public class Subject extends SaveState{
         }
 
 
+        if (success){//making the pdfs folder
+            imgDir.mkdir();
+
+            if (!imgDir.exists()){
+                System.out.println("Failed to create the image directory for " + name);
+                this.name = "";//using to represent failure
+                success = false;
+            }
+        }
+
+
 
 
         if (success){// succeeded in making all the folders
             this.cardPath = this.filePath + "/CueCards.json";
             this.notesPath = this.filePath + "/notes";
             this.pdfPath = this.filePath + "/pdfs";
+            this.imagePath = this.filePath + "/images";
 
             if (new File(this.cardPath).exists()) {//loading the existing cue cards
                 this.cueCardsList = Load(this.cardPath, CueCard.class);
@@ -106,9 +123,14 @@ public class Subject extends SaveState{
 
 
         //checking for invalid input
-        if (newName.isBlank()){newName = "new folder";}
+        if (newName.isBlank()){newName = defaultName;}
         if (newName.contains(SaveState.devFolder)){
             System.out.println("User is not allowed to add files to the dev folder");
+
+            try{AlertBox.display("Error changing the Subject's name","User is not allowed to turn dev file into a Subject");}
+            catch(ExceptionInInitializerError error){}//error happens when the front end isn't initialized (like when testing the backend)
+            catch(NoClassDefFoundError error){}
+
             return -2;
         }
 
@@ -125,6 +147,8 @@ public class Subject extends SaveState{
                 this.cardPath = this.filePath + "/CueCards.json";
                 this.notesPath = this.filePath + "/notes";
                 this.pdfPath = this.filePath + "/pdfs";
+                this.imagePath = this.filePath + "/images";
+
 
 
                 return 0;
@@ -175,16 +199,23 @@ public class Subject extends SaveState{
      * @param files files to be deleted
      */
     private static void RecursiveDelete(File[] files){
-        for (int x = 0; x < files.length; x++){
-            if(files[x].isDirectory()){
-                RecursiveDelete(files[x].listFiles());//delete contents of subdirectory
-            }
+        if (files != null) {
+            for (int x = 0; x < files.length; x++) {
+                if (files[x].isDirectory()) {
+                    RecursiveDelete(files[x].listFiles());//delete contents of subdirectory
+                }
 
-            files[x].delete();
+                files[x].delete();
+            }
         }
     }
 
 
+    /***
+     * gets the file path of a subject
+     * @return the file path of a subject
+     */
+    public String GetSubjectFilePath(){return this.filePath;}
 
 
 
@@ -500,6 +531,15 @@ public class Subject extends SaveState{
      * @return the pdfPath
      */
     public String GetPDFFilePath(){return this.pdfPath;}
+
+
+
+    /***
+     * gets the imagePath for the Subject
+     * @return the imagePath
+     */
+    public String GetImageFilePath(){return this.imagePath;}
+
 
     /***
      * gets all pdf files from a folder, removes the .pdf file extensions
